Override GMC montage code, bind previous montage position.
This is at least a temporary solution to ensure no need to modify GMC directly.
The caveat here is that we are not 100% guaranteed to stay in sync for motion
warping if MontageUpdate changes the current position but doesn't actually
utilize it yet; there shouldn't be a way for this to happen AT PRESENT,
but it's a limitation of the method.